Refactor Nix package generation
Previously we were using haskellPackageGen, which generates packages
from their source, but this is using import-from-derivation
feature in Nix, meaning that evaluation and realization mix and match.

Now we have a pkgs/generate.sh file that also doesn't need to specify
sha256 so bumping packages is easier.

Soon that should be all automated using stack2nix.