[DEVOPS-108] Supporting overriding monitor config per deployment.
By moving the `mkMonitor` call to the call-site in each deployment
configuration we can override the default settings for a given
monitor-deployment pair. For example, this inherits the CPU monitors
defaults while overriding the warning threshold.

    cpu = mkMonitor (cpu_monitor // {
      monitorOptions.thresholds.warning = "0.8";
    });

Although `mkMonitor` itself doesn't have type constraints, the
upstream nixops definition of a monitor does. I verified changing a
monitors name to a number, for instance, causes an error.