Update unpriv relays dynamic subsciptions
We changed the DNS policy mechanism so it behaves like the static routes
with the outer list being multiple subscriptions (valency) and the inner
being alternatives (backups) for each other.

The policy here is to use primary relays from the same region first, and
for fallbacks use relays (ordered randomly) from other regions. The
random ordering is to avoid concentrating load upon failures.