[DEVOPS-606] iohk-ops: add Buildkite support for installer upload
iohk-ops s3upload and find-installers will now recognize buildkite
entries in the GitHub commit status.

The Buildkite API is used to find the URL to download the
installer. Access to this API is authenticated and the token is stored
in a config file static/buildkite_token.

The version number used to rename the file is retrieved from
.buildkite/pipeline.yml.