[DEVOPS-639] recovery-service: Revert back to local terraform statefile
Due to the 3rd party support requirements, it would be best to revert
to managing credentials `terraform.tfvars`.

We also concluded that due to the unique security
constraints (separate AWS accounts per staging/production) and, again,
supporting the 3rd party, we won't be using an S3 bucket to manage the
state file.

Workspaces aren't necessary under this simplified model, assuming the
staging and production are deployed from different contexts with
distinct credentials and `terraform.tfvars` files.