infra: Provide minimal build environment for buildkite-agent
Provide git-lfs in the Buildkite environment so that LFS repos can be
cloned from GitHub.

Set the NIX_PATH for the builds to be what's running on the
buildkite-agent system. Mostly, our builds will be pinning the nixpkgs
version, so this is usually irrelevant. But sometimes it's helpful to
be able to use a nix package from the host system's package set. For
example, docker.