Adds initial graylog logging, journalbeat support
Integrates graylog on the monitoring server with journalbeat
support.  A security group is implementing allowing cluster
members to send journalbeat logs over port 5044 to the
monitoring server.

The graylog, mongoDB and elasticsearch components are
installed to the same monitoring server with this commit.

The journalbeat input on the graylog server requires
manual setup.  A future commit will address automation of this.