Add automated journalbeat input
A graylog-preload systemd service was added to the
monitoring server.  This service invokes a bash script
shortly after boot which makes calls to the graylog API
to preload a graylog "content pack" JSON blob.
The bash script will currently handle the initial
JSON loading, but needs to be further extended to handle
the use case of pushing a JSON update to a pre-existing
monitoring server.