Add content pack auto upgrade capability
The graylog-preload.service and calling script
have been upgraded to handle installing a different
content pack automatically when the id, version,
revision or comment changes.  The content pack
JSON name and comment keys (where applicable)
need to be "monitorContentPack" for the script logic
to work.

View the script called in the graylog-preload
service for more detail on functionality.