Graylog preload svc, script mods and Nginx changes
The graylog-preload systemd service was updated to call the content
pack auto-update script with a content_pack CLI argument.  The bash
script was also modified to accept this parameter.  This eliminated
the need to parse the shell script to inject Nix paths and allowed
for the script to be read directly into the Nix store.

The monitoring-services.nix file was updated to fix a graylog
swagger API incompatiblity with proxy_pass decoded URLs:

https://stackoverflow.com/q/28684300/

The root webhost URL was also updated to redirect to graylog by
default:

http(s)://$host/ -> https://$host/graylog